fix: UI, disable unreliable models

+++ src/contexts/ImageProcessingContext/ImageProcessingProvider.tsx
@@ -107,7 +107,7 @@ export const ImageProcessingProvider = ({
       const modelId = selectedModel;
 
       model.current = await AutoModel.from_pretrained(modelId, {
-        dtype: "fp32",
+        dtype: "fp16",
         //@ts-expect-error navigator gpu
         device: navigator?.gpu ? "webgpu" : "wasm",
         progress_callback: progresCallback,
